requirements.txt
astroid==2.0.4
autopep8==1.4
certifi==2018.8.24
chardet==3.0.4
click==6.7
colorama==0.3.9
feedparser==5.2.1
Flask==1.0.2
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.4.0
pylint==2.1.1
pywin32==223
requests==2.19.1
six==1.11.0
urllib3==1.23
virtualenv==16.0.0
weather-api==1.0.4
Werkzeug==0.14.1
wrapt==1.10.11
